undo moved code
